requirements.txt
hydra-core
tqdm
tensorflow-probability
wandb
omegaconf
flax
h5py
matplotlib